refactor: extract PV/ranking storage into store.ranking repo
Split first per-domain repository out of database.py (proof-of-pattern).
Move 5 PV/ranking functions (upsert_municipality_pv,
save_municipality_pv_panel, get_pv_profiles, get_pv_movers,
get_municipality_pv_panel) into store/ranking.py. They resolve the
connection seam via database.get_connection at call time, so existing
monkeypatches and callers (db.get_pv_profiles()) keep working.
database.py re-exports the functions; zero caller edits, zero behaviour
change. Add focused interface tests in tests/test_store_ranking.py.
